refactor: move cart and nav open states
