+++ layouts/default.vue
@@ -15,13 +15,7 @@
 </template>
 
 <script>
-import {
-  inject,
-  provide,
-  ref,
-  useContext,
-  watch,
-} from "@nuxtjs/composition-api";
+import { inject } from "@nuxtjs/composition-api";
 
 import { SpaceProvider, EventProvider, CartProvider } from "@nacelle/vue";
 import SiteContainer from "~/components/core/SiteContainer.vue";
@@ -34,24 +28,8 @@ export default {
     SiteContainer,
   },
   setup() {
-    const cartOpen = ref(false);
-    const navOpen = ref(false);
     const initialSpace = inject("initialSpace");
     const nacelleSdk = inject("nacelleSdk");
-    const { route } = useContext();
-
-    const setCartOpen = (val) => (cartOpen.value = val);
-    const setNavOpen = (val) => (navOpen.value = val);
-
-    watch(route, () => {
-      cartOpen.value = false;
-      navOpen.value = false;
-    });
-
-    provide("cartOpen", cartOpen);
-    provide("navOpen", navOpen);
-    provide("setCartOpen", setCartOpen);
-    provide("setNavOpen", setNavOpen);
 
     return { initialSpace, nacelleSdk };
   },
